Move flagsaver to testonly package tree.
This code is only intended to be used by tests.
